requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
burnysc2==6.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.7
loguru==0.6.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mpyq==0.2.5
multidict==6.1.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.2
portpicker==1.6.0
protobuf==3.20.3
psutil==6.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s2clientprotocol==5.0.13.92440.3
scipy==1.14.1
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==74.1.2
Shimmy==0.2.1
six==1.16.0
smmap==5.0.1
stable_baselines3==2.3.2
sympy==1.13.2
tensorflow
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==2.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.0
Werkzeug==3.0.4
yarl==1.11.1